pyproject.toml
[project]
name = "ogh-ss25-hackathon-canopy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "geopandas>=1.1.1",
    "jupyterlab>=4.4.7",
    "matplotlib>=3.10.6",
    "numpy>=2.3.2",
    "scikit-learn>=1.7.1",
    "torch>=2.8.0",
]
